post-refactor tuning N+1
