+++ src/scripts/editorUtils.js
@@ -86,6 +86,133 @@ export function isEditableDiv(elem) {
     return true;
 }
 
+export function marcFieldToDiv(recordDiv, originalRow = undefined, field, settings = null, fieldIsEditable = false, altDocument = undefined) {
+  // This function should not be used imported (expect for testing)!
+
+  // altDocument is a jsdom document. We need it for testing.
+  const myDocument = altDocument || document;
+
+  // NB! Typically recordDiv (parent of row div) or originalRow div is empty, as only one of them is needed.
+  const row = originalRow || myDocument.createElement('div');
+  row.classList.add('row');
+
+  if (field.uuid) {
+    row.id = field.uuid;
+  }
+
+  if (settings?.decorateField) {
+    settings.decorateField(row, field);
+  }
+  if (settings?.onClick) { // add similar keydown or input event?
+    row.addEventListener('click', event => settings.onClick(event, field));
+  }
+
+  if (fieldIsEditable) {
+    row.setAttribute('contentEditable', true);
+  }
+  else {
+    if (settings?.uneditableFieldBackgroundColor) {
+      row.style.backgroundColor = settings.uneditableFieldBackgroundColor;
+    }
+  }
+
+  addTag(row, field.tag);
+
+  // NB! Note that the current implementation will add two non-breaking spaces for indicatorless fields.
+  addIndicators(row);
+
+  if (field.subfields) {
+    for (const [index, subfield] of field.subfields.entries()) {
+      addSubfield(row, subfield, index);
+    }
+  } else {
+    addValue(row, field.value);
+  }
+
+  if (recordDiv && !originalRow) {
+    recordDiv.appendChild(row);
+  }
+  return row;
+
+  //---------------------------------------------------------------------------
+
+  function addTag(row, value) {
+    row.appendChild(makeSpan('tag', value));
+  }
+
+  function addIndicators(row) {
+    const span = makeSpan('inds');
+    if (field.ind1 || field.value) { // Field in editor might be incomplete and lack indicators
+      addIndicator(span, field.ind1, 'ind1');
+      if (field.ind2 || field.value) {
+        addIndicator(span, field.ind2, 'ind2');
+      }
+    }
+    row.appendChild(span);
+
+    function addIndicator(span, ind, className = 'ind') {
+      // Rather hackily a <span class="${className}">&nbsp;</span> is created for non-indicator fields...
+      const value = mapIndicatorToValue(ind);
+      span.appendChild(makeSpan(className, null, value));
+    }
+
+    function mapIndicatorToValue(ind) { // field -> web page
+      if (!isDataFieldTag(field.tag)) {
+        return '&nbsp;'; // ' ' or &nbsp;?
+      }
+      if ( ind === ' ') {
+        return '#'; // '#' is the standard way to represent an empty indicator.
+      }
+      return ind;
+    }
+  }
+
+  function addValue(row, value) {
+    const normalizedValue  = normalizeValue();
+    row.appendChild(makeSpan('value', normalizedValue));
+
+    function normalizeValue() {
+      if (!value) {
+        return '';
+      }
+      if (!isDataFieldTag(field.tag)) {
+        return field.value.replace(/ /gu, '#');
+      }
+      return value;
+    }
+  }
+
+  function addSubfield(row, subfield, index = 0) {
+    const span = makeSpan('subfield');
+    span.appendChild(makeSubfieldCode(`${subfield.code}`, index));
+    span.appendChild(makeSubfieldData(subfield.value, index));
+    row.appendChild(span);
+  }
+
+  function makeSubfieldCode(code, index = 0) {
+    if (settings?.subfieldCodePrefix) {
+      return makeSpan('code', `${settings.subfieldCodePrefix}${code}`, null, index);
+    }
+    return makeSpan('code', code, null, index);
+  }
+
+  function makeSubfieldData(value, index = 0) {
+    return makeSpan('value', value, null, index);
+  }
+
+  //-----------------------------------------------------------------------------
+  function makeSpan(className, text, html, index = 0) {
+    const span = myDocument.createElement('span');
+    span.setAttribute('class', className);
+    span.setAttribute('index', index);
+    if (text) {
+      span.textContent = text;
+    } else if (html) {
+      span.innerHTML = html;
+    }
+    return span;
+  }
+}
 
 export function markAllFieldsUneditable(settings) {
     // I call this typically after clicking save button.
